feat(ui): type bounded scalar controls
Introduce canonical Zoom, TempoPercent, and FaderValue domains. Carry the fader coordinate through the player-to-host slot boundary while preserving unbounded HostLevel amplitude semantics.

+++ crates/kithara-app/src/gui/ui/cache.rs
@@ -278,7 +278,7 @@ impl ViewCache {
 impl DeckCache {
     fn refresh(&mut self, deck: &DeckUi) {
         let ts = deck.view.timestretch;
-        self.tempo = format!("{:+.1}%", ts.tempo);
+        self.tempo = format!("{:+.1}%", f32::from(ts.tempo));
         self.bpm = format_bpm(analysis_bpm(&deck.ui), ts.speed());
         self.remain = format_remain(&deck.ui);
         self.subtitle = track_subtitle(&deck.ui);
